feat: code quality cleanup — deduplicate, fix bare excepts, fix docs
- Extract 7 duplicated functions to tools/shared/media_utils.py
- Extract strip_track_number to tools/shared/text_utils.py
- fix_dynamic_track imports apply_eq/soft_clip from master_tracks (gains safety guards)
- Replace 6 bare except: clauses with specific exception types
- Guard import yaml in sheet-music tools to match project convention
- Remove unused import (ProgressBar), dead code (BG_COLOR, WAVEFORM_COLOR, mask)
- Fix Image.open resource leak in album sampler
- Fix PR template Co-Authored-By (Sonnet → Opus 4.5)
- Fix stale skill counts in model-strategy.md and SKILL_INDEX.md
- Fix README stale paths.yaml reference, false content_root default claim
- Add SKILL_INDEX.md and model-strategy.md to CONTRIBUTING new skill checklist
- Clarify version bumps happen at release time in CONTRIBUTING
