+++ .claude-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "bitwize-music",
   "description": "AI music generation workflow for Suno - album concepts, lyrics, prompts, mastering, release",
-  "version": "0.34.1",
+  "version": "0.35.0",
   "author": {
     "name": "bitwize-music"
   },
